+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@
 
 Issue #14654 - Add processing of interp_lat,lon parameters
 
+Issue #14791 - Correct pressure_depth/int_ctd_pressure handling
+
 Issue #14790 - Add support for QARTOD climatology tests
 
 # Release 1.15.1 2020-06-12